Split public and internal disposable stuff.
